Use int64

+++ ratelimit.go
@@ -8,23 +8,23 @@ import (
 
 type RateLimiter interface {
 	Start()
-	GetMetrics() map[string]uint
+	GetMetrics() map[string]int64
 	Read(p []byte) (n int, err error)
 	Write(p []byte) (n int, err error)
 }
 
 type rateLimiter struct {
-	droppedBytes    uint
-	droppedMessages uint
-	successBytes    uint
-	successMessages uint
+	droppedBytes    int64
+	droppedMessages int64
+	successBytes    int64
+	successMessages int64
 	inChan          chan []byte
 	outChan         chan []byte
 	limiter         *rate.Limiter
 }
 
-func (rl *rateLimiter) GetMetrics() map[string]uint {
-	return map[string]uint{
+func (rl *rateLimiter) GetMetrics() map[string]int64 {
+	return map[string]int64{
 		"droppedBytes":    rl.droppedBytes,
 		"droppedMessages": rl.droppedMessages,
 		"successBytes":    rl.successBytes,
@@ -57,10 +57,10 @@ func (rl *rateLimiter) Write(p []byte) (int, error) {
 	select {
 	case rl.inChan <- p:
 		rl.successMessages = rl.successMessages + 1
-		rl.successBytes = rl.successBytes + uint(len(p))
+		rl.successBytes = rl.successBytes + int64(len(p))
 	default:
 		rl.droppedMessages = rl.droppedMessages + 1
-		rl.droppedBytes = rl.droppedBytes + uint(len(p))
+		rl.droppedBytes = rl.droppedBytes + int64(len(p))
 	}
 	return len(p), nil
 }
